chore: update testing
